requirements.txt
PyPDF2
openai
fastapi
uvicorn
python-multipart
python-dotenv